+++ CHANGELOG.md
@@ -3,11 +3,16 @@ Unreleased
 
 ## Added:
 
+ * Support for nonbreaking space character (`\u{A0}`).
  * Added optimized measurement function that do not expect carriage returns.
    * `FontExt::measure_line_nocr`
    * `FontExt::str_width_nocr`
    * `FontExt::max_str_width_nocr`
  * Support carriage return (`\r`) control characters.
+
+## Fixed:
+
+ * Fixed an issue where height measurement unexpectedly carried a space that is consumed during drawing.
 
 0.1.0 (2020-07-31)
 ==================